reorg patches
