+++ SolastaCommunityExpansion/Patches/SrdAndHouseRules/ExperienceMultiply/RulesetCharacterHeroPatcher.cs
@@ -4,23 +4,8 @@ using System.Linq;
 using HarmonyLib;
 using static SolastaModApi.DatabaseHelper.QuestTreeDefinitions;
 
-namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules
+namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules.ExperienceMultiply
 {
-    [HarmonyPatch(typeof(RulesetCharacterHero), "GrantItem")]
-    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
-    internal static class RulesetCharacterHero_GrantItem
-    {
-        public static void Prefix(ref bool tryToEquip)
-        {
-            if (Main.Settings.DisableAutoEquip)
-            {
-                var characterBuildingService = ServiceRepository.GetService<ICharacterBuildingService>();
-
-                tryToEquip = characterBuildingService?.HeroCharacter != null;
-            }
-        }
-    }
-
     [HarmonyPatch(typeof(RulesetCharacterHero), "GrantExperience")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class RulesetCharacterHero_GrantExperience
