+++ SolastaCommunityExpansion/Patches/SrdAndHouseRules/OfficialCombatSurpriseRules/GameLocationBattlePatcher.cs
@@ -1,7 +1,7 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
 
-namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules
+namespace SolastaCommunityExpansion.Patches.SrdAndHouseRules.OfficialCombatSurpriseRules
 {
     [HarmonyPatch(typeof(GameLocationBattle), "StartContenders")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
